+++ negotiation/src/main/java/it/eng/negotiation/serializer/Serializer.java
@@ -62,9 +62,6 @@ public class Serializer {
         };
         
 		jsonMapperPlain = JsonMapper.builder()
-//				.configure(MapperFeature.USE_ANNOTATIONS, false)
-//				.serializationInclusion(Include.NON_NULL)
-//				.serializationInclusion(Include.NON_EMPTY)
 				.configure(SerializationFeature.INDENT_OUTPUT, true)
 				.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false)
 				.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
@@ -76,7 +73,6 @@ public class Serializer {
 				.serializationInclusion(Include.NON_EMPTY)
 				.configure(SerializationFeature.INDENT_OUTPUT, true)
 				.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false)
-//			.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
 				.build();
 		
 		validator = Validation.buildDefaultValidatorFactory().getValidator();