+++ CHANGELOG.md
@@ -14,6 +14,10 @@ All notable changes to this project will be documented in this file.
  
 ## [0.0.1] - xx-09-2024
 
+### Added
+
+ - GHA test for API endpoints
+ 
 ### Changed
 
  - Plain serializers returns '@id'