+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.0] - 2024-08-xx
+## [0.1.0] - 13-09-2024
 
 ### Added
  
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
  - Postman collection for testing endpoints	
  - Configured GitHub actions to run tests
  
-## [0.0.1] - xx-09-2024
+## [0.0.1] - 12-09-2024
 
 ### Added